+++ apipgen.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'apipgen'
 
 const config = defineConfig({
-  pipeline: 'swag-js-fetch',
+  pipeline: 'swag-js-ky',
   input: {
     uri: 'https://petstore.swagger.io/v2/swagger.json',
   },