+++ packages/swag-js-ky/src/config/index.ts
@@ -1,0 +1,18 @@
+import type { ApiPipeline } from 'apipgen'
+import { readConfig as _readConfig } from '@apipgen/pipeline'
+
+export function readConfig(config: ApiPipeline.Config): ApiPipeline.ConfigRead {
+  config.import = config.import || {}
+  config.output = config.output || {}
+  config.import.http = config.import.http || 'ky'
+  config.output.main = config.output.main?.replace('.ts', '.js') || 'src/api/index.js'
+
+  const configRead = _readConfig(config)
+
+  configRead.graphs.imports.push({
+    name: 'http',
+    value: config.import.http,
+  })
+
+  return configRead
+}