+++ packages/swag-js-fetch/CHANGELOG.md
@@ -1,5 +1,13 @@
 # apipgen-swag-js-fetch
 
+## 0.1.1
+
+### Patch Changes
+
+- feat: add swag-ky support
+  - @apipgen/swag-parser@0.1.0
+  - @apipgen/pipeline@0.1.0
+
 ## 0.1.0
 
 ### Minor Changes